requirements.txt
requests
SQLAlchemy
alembic
python-dotenv